Refactor so that tests need only import azure_core_tests
Also makes the attribute `#[recorded::test]` which, IMO, looks a bit better.
