+++ sdk/core/azure_core_test_macros/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "azure_core_macros"
+name = "azure_core_test_macros"
 version = "0.1.0"
-description = "Procedural macros for client libraries built on azure_core."
+description = "Procedural macros for testing client libraries built on azure_core."
 readme = "README.md"
 authors.workspace = true
 license.workspace = true
@@ -13,17 +13,15 @@ categories = ["development-tools"]
 edition.workspace = true
 rust-version.workspace = true
 
+[lib]
+proc-macro = true
+
 [dependencies]
-typespec = { workspace = true, features = ["http", "json"] }
-typespec_client_core = { workspace = true, features = ["http", "json"] }
-azure_core_test.workspace = true
+azure_core = { workspace = true, features = ["test"] }
 proc-macro2.workspace = true
 quote.workspace = true
 syn.workspace = true
 
 [dev-dependencies]
-azure_core.workspace = true
+azure_core_test.workspace = true
 tokio.workspace = true
-
-[lib]
-proc-macro = true
