+++ sdk/core/azure_core_test_macros/README.md
@@ -5,20 +5,19 @@ Macros for testing client libraries built on `azure_core`.
 ## Client methods
 
 To test client methods using our [Test Proxy], you can attribute both synchronous and asynchronous (recommend) tests
-using the `#[recorded]` attribute:
+using the `#[recorded::test]` attribute:
 
 ```rust
-use azure_core_macros::recorded;
-use azure_core_test::TestContext;
+use azure_core_test::{recorded, TestContext};
 use azure_core::Result;
 
-#[recorded]
+#[recorded::test]
 async fn get_secret(ctx: TestContext) -> Result<()> {
     todo!()
 }
 ```
 
-The `TestContext` parameter is required unless your test function is attribute as `#[recorded(live)]` (live-only),
+The `TestContext` parameter is required unless your test function is attribute as `#[recorded::test(live)]` (live-only),
 in which case it is optional. You can name the parameter whatever you want.
 The `TestContext` parameter is used to initialize an HTTP client to play back or record tests
 and provides other information to test functions that may be useful.
