+++ sdk/cosmos/azure_data_cosmos/src/clients/cosmos_client.rs
@@ -69,7 +69,7 @@ impl CosmosClient {
     /// ```rust,no_run
     /// use azure_data_cosmos::CosmosClient;
     ///
-    /// let client = CosmosClient::with_key("https://myaccount.documents.azure.com/", "my_key", None).unwrap();
+    /// let client = CosmosClient::with_key("https://myaccount.documents.azure.com/", "my_key".into(), None).unwrap();
     /// ```
     #[cfg(feature = "key_auth")]
     pub fn with_key(
