+++ sdk/eventhubs/azure_messaging_eventhubs/Cargo.toml
@@ -32,7 +32,6 @@ rustc_version.workspace = true
 
 [dev-dependencies]
 tracing-subscriber = { workspace = true, features = ["env-filter", "fmt"] }
-azure_core_macros.workspace = true
 azure_core_test.workspace = true
 azure_identity.workspace = true
 tokio = { workspace = true, default-features = false, features = [
