+++ sdk/eventhubs/azure_messaging_eventhubs/tests/producer.rs
@@ -7,7 +7,7 @@ use azure_core_amqp::{
     messaging::{AmqpMessage, AmqpMessageBody},
     value::{AmqpList, AmqpValue},
 };
-use azure_core_macros::recorded;
+use azure_core_test::recorded;
 use azure_identity::DefaultAzureCredential;
 use azure_messaging_eventhubs::producer::{
     batch::EventDataBatchOptions, ProducerClient, ProducerClientOptions,
@@ -17,7 +17,7 @@ use tracing::{info, trace};
 
 mod common;
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_new() {
     common::setup();
     let host = env::var("EVENTHUBS_HOST").unwrap();
@@ -33,7 +33,7 @@ async fn test_new() {
     );
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_new_with_error() {
     common::setup();
     let eventhub = env::var("EVENTHUB_NAME").unwrap();
@@ -51,7 +51,7 @@ async fn test_new_with_error() {
     info!("Error: {:?}", result);
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_open() {
     common::setup();
     let host = env::var("EVENTHUBS_HOST").unwrap();
@@ -67,7 +67,7 @@ async fn test_open() {
     );
     client.open().await.unwrap();
 }
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_close() {
     common::setup();
     let host = env::var("EVENTHUBS_HOST").unwrap();
@@ -85,7 +85,7 @@ async fn test_close() {
     client.close().await.unwrap();
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_get_properties() {
     common::setup();
     let host = env::var("EVENTHUBS_HOST").unwrap();
@@ -108,7 +108,7 @@ async fn test_get_properties() {
     assert_eq!(properties.name, eventhub);
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_get_partition_properties() {
     common::setup();
     let host = env::var("EVENTHUBS_HOST").unwrap();
@@ -138,7 +138,7 @@ async fn test_get_partition_properties() {
     }
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 fn test_create_eventdata() {
     common::setup();
     let data = b"hello world";
@@ -162,7 +162,7 @@ fn test_create_eventdata() {
         .build();
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_create_batch() {
     common::setup();
     let host = env::var("EVENTHUBS_HOST").unwrap();
@@ -186,7 +186,7 @@ async fn test_create_batch() {
     }
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_create_and_send_batch() {
     common::setup();
     let host = env::var("EVENTHUBS_HOST").unwrap();
@@ -238,7 +238,7 @@ async fn test_create_and_send_batch() {
     }
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_add_amqp_messages_to_batch() -> Result<(), Box<dyn std::error::Error>> {
     common::setup();
     let host = env::var("EVENTHUBS_HOST").unwrap();
@@ -315,7 +315,7 @@ async fn test_add_amqp_messages_to_batch() -> Result<(), Box<dyn std::error::Err
     Ok(())
 }
 
-#[recorded(live)]
+#[recorded::test(live)]
 async fn test_overload_batch() {
     common::setup();
